# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evolution API Proxy",
    "version": "1.0.0",
    "description": "Proxy backend for Evolution API enabling ChatGPT Actions integration for WhatsApp automation",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-production-domain.com/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Evolution API key for authentication"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateInstanceRequest": {
        "type": "object",
        "required": ["instanceName"],
        "properties": {
          "instanceName": {
            "type": "string",
            "description": "Unique name for the WhatsApp instance",
            "example": "my-whatsapp-bot"
          },
          "qrcode": {
            "type": "boolean",
            "description": "Generate QR code for connection",
            "example": true
          },
          "integration": {
            "type": "string",
            "description": "Integration type",
            "example": "WHATSAPP-BAILEYS"
          }
        }
      },
...
      "CreateGroupRequest": {
        "type": "object",
        "required": ["subject", "participants"],
        "properties": {
          "subject": {
            "type": "string",
            "description": "Group name",
            "example": "My WhatsApp Group"
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of phone numbers to add",
            "example": ["5511999999999", "5511888888888"]
          }
        }
      },
      "UpdateGroupParticipantsRequest": {
        "type": "object",
        "required": ["groupJid", "action", "participants"],
        "properties": {
          "groupJid": {
            "type": "string",
            "description": "Group JID",
            "example": "123456789@g.us"
          },
          "action": {
            "type": "string",
            "enum": ["add", "remove", "promote", "demote"],
            "description": "Action to perform on participants",
            "example": "add"
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["5511999999999"]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/instances": {
      "get": {
        "summary": "List all WhatsApp instances",
        "description": "Retrieve a list of all WhatsApp instances",
        "operationId": "listInstances",
        "tags": ["Instance Management"],
        "responses": {
          "200": {
            "description": "List of instances",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new WhatsApp instance",
        "description": "Create a new WhatsApp instance with a unique name",
        "operationId": "createInstance",
        "tags": ["Instance Management"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstanceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instance created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}": {
      "get": {
        "summary": "Get instance details",
        "description": "Get detailed information about a specific instance",
        "operationId": "getInstanceDetails",
        "tags": ["Instance Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the instance"
          }
        ],
        "responses": {
          "200": {
            "description": "Instance details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Instance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete instance",
        "description": "Delete a WhatsApp instance",
        "operationId": "deleteInstance",
        "tags": ["Instance Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instance deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/status": {
      "get": {
        "summary": "Get instance connection status",
        "description": "Get the current connection status of an instance",
        "operationId": "getInstanceStatus",
        "tags": ["Instance Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instance status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/connect": {
      "post": {
        "summary": "Connect instance",
        "description": "Connect instance and get QR code for pairing",
        "operationId": "connectInstance",
        "tags": ["Instance Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection initiated, QR code returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/disconnect": {
      "post": {
        "summary": "Disconnect instance",
        "description": "Disconnect and logout from WhatsApp instance",
        "operationId": "disconnectInstance",
        "tags": ["Instance Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instance disconnected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/instances/{instanceName}/contacts": {
      "get": {
        "summary": "List all contacts",
        "description": "Get list of all contacts for an instance",
        "operationId": "listContacts",
        "tags": ["Contact Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/contacts/{contactId}": {
      "get": {
        "summary": "Get contact details",
        "description": "Get details of a specific contact",
        "operationId": "getContactDetails",
        "tags": ["Contact Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/instances/{instanceName}/groups": {
      "get": {
        "summary": "List all groups",
        "description": "Get list of all WhatsApp groups",
        "operationId": "listGroups",
        "tags": ["Group Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new group",
        "description": "Create a new WhatsApp group",
        "operationId": "createGroup",
        "tags": ["Group Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/groups/{groupId}": {
      "get": {
        "summary": "Get group details",
        "description": "Get details of a specific group",
        "operationId": "getGroupDetails",
        "tags": ["Group Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/groups/{groupId}/participants": {
      "post": {
        "summary": "Manage group participants",
        "description": "Add, remove, promote, or demote group participants",
        "operationId": "updateGroupParticipants",
        "tags": ["Group Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupParticipantsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Participants updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/instances/{instanceName}/chats": {
      "get": {
        "summary": "List all chats",
        "description": "Get list of all chats for instance",
        "operationId": "listChats",
        "tags": ["Chat Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of chats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/chats/{chatId}": {
      "get": {
        "summary": "Get chat details",
        "description": "Get details of a specific chat",
        "operationId": "getChatDetails",
        "tags": ["Chat Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete chat",
        "description": "Delete a chat conversation",
        "operationId": "deleteChat",
        "tags": ["Chat Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}